requirements.txt
flask
line-bot-sdk
apscheduler
pytz
google-generativeai
gunicorn
feedparser
yfinance